Split `layout_2020/flexbox.rs` into modules
